dtk_lib/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
boto3==1.37.4
